+++ src/Geometry.cpp
@@ -6,18 +6,18 @@ bool Sphere::hit(const Ray& ray, const Interval& ray_t, HitRecord& record) const
     Vec3 dir = ray.GetDirection();
 
     // h == b/-2 allowing a simplification of the quadratic
-    double a{ dir.LengthSquared() };
-    double h{ Dot(dir, oc) };
-    double c{ oc.LengthSquared() - m_radius * m_radius };
+    float a{ dir.LengthSquared() };
+    float h{ Dot(dir, oc) };
+    float c{ oc.LengthSquared() - m_radius * m_radius };
 
-    double discriminant{ h * h - a * c };
+    float discriminant{ h * h - a * c };
     if (discriminant < 0) {
         return false;
     }
 
-    double sqrt_d{ std::sqrt(discriminant) };
+    float sqrt_d{ std::sqrt(discriminant) };
 
-    double root = (h - sqrt_d) / a;
+    float root = (h - sqrt_d) / a;
     if (!ray_t.Surrounds(root)) {
         root = (h + sqrt_d) / a;
         if (!ray_t.Surrounds(root)) {
