+++ src/Interval.h
@@ -5,31 +5,31 @@
 class Interval {
 public:
     Interval() : m_min(+infinity), m_max(-infinity) {} // empty interval
-    Interval(double min, double max) : m_min(min), m_max(max) {}
+    Interval(float min, float max) : m_min(min), m_max(max) {}
 
-    bool Contains(double x) const {
+    bool Contains(float x) const {
         return m_min <= x && x <= m_max;
     }
 
-    bool Surrounds(double x) const {
+    bool Surrounds(float x) const {
         return m_min < x && x < m_max;
     }
 
-    double Clamp(double x) const {
+    float Clamp(float x) const {
         if (x < m_min) return m_min;
         if (x > m_max) return m_max;
         return x;
     }
 
-    double Min() const {
+    float Min() const {
         return m_min;
     }
 
-    double Max() const {
+    float Max() const {
         return m_max;
     }
 
-    double Size() const {
+    float Size() const {
         return m_max - m_min;
     }
 
@@ -42,6 +42,6 @@ public:
     }
 
 private:
-    double m_min{};
-    double m_max{};
+    float m_min{};
+    float m_max{};
 };
